More

+++ src/ML/MLOptimiser.cpp
@@ -1473,12 +1473,12 @@ double logDataVSPrior(const Image& dat,
 {
     double result = 0;
 
-    int r2 = gsl_pow_2(r);
-    int d2 = gsl_pow_2(FREQ_DOWN_CUTOFF);
+    double r2 = gsl_pow_2(r);
+    double d2 = gsl_pow_2(FREQ_DOWN_CUTOFF);
 
     IMAGE_FOR_PIXEL_R_FT(r + 1)
     {
-        int u = QUAD(i, j);
+        double u = QUAD(i, j);
 
         if ((u < r2) && (u > d2))
         {
@@ -1512,7 +1512,7 @@ double logDataVSPrior(const Image& dat,
 
     IMAGE_FOR_PIXEL_R_FT(r + 1)
     {
-        int u = QUAD(i, j);
+        double u = QUAD(i, j);
 
         if ((u < r2) && (u > d2))
         {
